:white_check_mark: Shorten basic example log message to fix tests
In Python 3.7, an expression's canonical line number is the one which
corresponds to the *last* line of the expression, as opposed to Python
>=3.8 where the line number corresponding to the *first* line of an
expression is that expression's canonical line number.

E.g., prior to this change, the line number of the log message emitted
in `pure_structlog_logging_without_sentry.py` was 6 in Python 3.7 (the
line corresponding to the logger arguments) and 5 in Python 3.8 & 3.9
(the line corresponding to the logger invocation).