:bug: Bind logger name to meta logger
This is also added manually by the library for user-requested loggers,
so not adding this prior was a bug.

This also resolves the other, more serious bug where the logger would
not honor the requested minimum log level or local development
formatting settings:

Previously initialization was deferred and these configs weren't being
propagated to the meta logger on that initialization. Binding the logger
 with a logger name initializes the logger with the configs (before they
can be reset) and caches the logger future usage, ensuring the
requested log level is honored, even for the library meta logger.